HDX-9781 Move operation_presence view as table
