+++ src/hapi_schema/db_views_as_tables.py
@@ -177,3 +177,30 @@ class DBNationalRiskVAT(Base):
     )
     location_code: Mapped[str] = mapped_column(String(128), index=True)
     location_name: Mapped[str] = mapped_column(String(512), index=True)
+
+
+class DBOperationalPresenceVAT(Base):
+    __tablename__ = "operational_presence_vat"
+    resource_hdx_id: Mapped[str] = mapped_column(String(36))
+    admin2_ref: Mapped[int] = mapped_column(Integer, primary_key=True)
+    org_acronym: Mapped[str] = mapped_column(String, primary_key=True)
+    org_name: Mapped[str] = mapped_column(String, primary_key=True)
+    sector_code: Mapped[str] = mapped_column(String(32), primary_key=True)
+    reference_period_start: Mapped[datetime] = mapped_column(
+        DateTime, primary_key=True
+    )
+    reference_period_end: Mapped[datetime] = mapped_column(
+        DateTime, index=True
+    )
+    org_type_code: Mapped[str] = mapped_column(String(32))
+    sector_name: Mapped[str] = mapped_column(String(512))
+    location_code: Mapped[str] = mapped_column(String(128), index=True)
+    location_name: Mapped[str] = mapped_column(String(512), index=True)
+    admin1_code: Mapped[str] = mapped_column(String(128), index=True)
+    admin1_name: Mapped[str] = mapped_column(String(512), index=True)
+    admin1_is_unspecified: Mapped[bool] = mapped_column(Boolean)
+    location_ref: Mapped[int] = mapped_column(Integer)
+    admin2_code: Mapped[str] = mapped_column(String(128), index=True)
+    admin2_name: Mapped[str] = mapped_column(String(512), index=True)
+    admin2_is_unspecified: Mapped[bool] = mapped_column(Boolean)
+    admin1_ref: Mapped[int] = mapped_column(Integer)
